backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-nmap==0.7.1
celery==5.3.6
redis==5.0.1
httpx==0.26.0
google-generativeai==0.3.2
jinja2==3.1.3
python-multipart==0.0.6
pydantic-settings==2.1.0
requests==2.31.0
pytest==8.0.0
reportlab==4.0.7
